impl zuc256 without mac generator which pass basic test (#3)
