+++ README.md
@@ -11,3 +11,7 @@
 [License]: https://img.shields.io/crates/l/zuc.svg
 
 Documentation: <https://docs.rs/zuc>
+
+## References
+- **ZUC 128**:[GB/T 33133.1-2016](https://openstd.samr.gov.cn/bzgk/gb/newGbInfo?hcno=8C41A3AEECCA52B5C0011C8010CF0715)
+- **ZUC 256**:[ZUC256-version1.1](http://www.is.cas.cn/ztzl2016/zouchongzhi/201801/W020180416526664982687.pdf)
