+++ src/lib.rs
@@ -11,5 +11,12 @@
 // ---
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
+mod zuc;
+pub use self::zuc::ZUC;
 mod zuc128;
 pub use self::zuc128::ZUC128;
+mod zuc256;
+pub use self::zuc256::{MacLength, ZUC256};
+
+mod utils;
+mod zuc_data;
