+++ src/zuc_data.rs
@@ -1,0 +1,76 @@
+/// d constants 128
+pub static D_128: [u16; 16] = [
+    0b_0100_0100_1101_0111,
+    0b_0010_0110_1011_1100,
+    0b_0110_0010_0110_1011,
+    0b_0001_0011_0101_1110,
+    0b_0101_0111_1000_1001,
+    0b_0011_0101_1110_0010,
+    0b_0111_0001_0011_0101,
+    0b_0000_1001_1010_1111,
+    0b_0100_1101_0111_1000,
+    0b_0010_1111_0001_0011,
+    0b_0110_1011_1100_0100,
+    0b_0001_1010_1111_0001,
+    0b_0101_1110_0010_0110,
+    0b_0011_1100_0100_1101,
+    0b_0111_1000_1001_1010,
+    0b_0100_0111_1010_1100,
+];
+/// d constants 128
+pub static D_256: [u8; 16] = [
+    0b010_0010, 0b010_1111, 0b010_0100, 0b010_1010, 0b110_1101, 0b100_0000, 0b100_0000, 0b100_0000,
+    0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b101_0010, 0b001_0000, 0b011_0000,
+];
+pub static MAC_256_32: [u8; 16] = [
+    0b010_0010, 0b010_1111, 0b010_0101, 0b010_1010, 0b110_1101, 0b100_0000, 0b100_0000, 0b100_0000,
+    0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b101_0010, 0b001_0000, 0b011_0000,
+];
+pub static MAC_256_64: [u8; 16] = [
+    0b010_0011, 0b010_1111, 0b010_0100, 0b010_1010, 0b110_1101, 0b100_0000, 0b100_0000, 0b100_0000,
+    0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b101_0010, 0b001_0000, 0b011_0000,
+];
+pub static MAC_256_128: [u8; 16] = [
+    0b010_0011, 0b010_1111, 0b010_0101, 0b010_1010, 0b110_1101, 0b100_0000, 0b100_0000, 0b100_0000,
+    0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b100_0000, 0b101_0010, 0b001_0000, 0b011_0000,
+];
+
+/// S0 box
+pub static S0: [u8; 256] = const_str::hex!([
+    "3E 72 5B 47 CA E0 00 33 04 D1 54 98 09 B9 6D CB",
+    "7B 1B F9 32 AF 9D 6A A5 B8 2D FC 1D 08 53 03 90",
+    "4D 4E 84 99 E4 CE D9 91 DD B6 85 48 8B 29 6E AC",
+    "CD C1 F8 1E 73 43 69 C6 B5 BD FD 39 63 20 D4 38",
+    "76 7D B2 A7 CF ED 57 C5 F3 2C BB 14 21 06 55 9B",
+    "E3 EF 5E 31 4F 7F 5A A4 0D 82 51 49 5F BA 58 1C",
+    "4A 16 D5 17 A8 92 24 1F 8C FF D8 AE 2E 01 D3 AD",
+    "3B 4B DA 46 EB C9 DE 9A 8F 87 D7 3A 80 6F 2F C8",
+    "B1 B4 37 F7 0A 22 13 28 7C CC 3C 89 C7 C3 96 56",
+    "07 BF 7E F0 0B 2B 97 52 35 41 79 61 A6 4C 10 FE",
+    "BC 26 95 88 8A B0 A3 FB C0 18 94 F2 E1 E5 E9 5D",
+    "D0 DC 11 66 64 5C EC 59 42 75 12 F5 74 9C AA 23",
+    "0E 86 AB BE 2A 02 E7 67 E6 44 A2 6C C2 93 9F F1",
+    "F6 FA 36 D2 50 68 9E 62 71 15 3D D6 40 C4 E2 0F",
+    "8E 83 77 6B 25 05 3F 0C 30 EA 70 B7 A1 E8 A9 65",
+    "8D 27 1A DB 81 B3 A0 F4 45 7A 19 DF EE 78 34 60",
+]);
+
+/// S1 box
+pub static S1: [u8; 256] = const_str::hex!([
+    "55 C2 63 71 3B C8 47 86 9F 3C DA 5B 29 AA FD 77",
+    "8C C5 94 0C A6 1A 13 00 E3 A8 16 72 40 F9 F8 42",
+    "44 26 68 96 81 D9 45 3E 10 76 C6 A7 8B 39 43 E1",
+    "3A B5 56 2A C0 6D B3 05 22 66 BF DC 0B FA 62 48",
+    "DD 20 11 06 36 C9 C1 CF F6 27 52 BB 69 F5 D4 87",
+    "7F 84 4C D2 9C 57 A4 BC 4F 9A DF FE D6 8D 7A EB",
+    "2B 53 D8 5C A1 14 17 FB 23 D5 7D 30 67 73 08 09",
+    "EE B7 70 3F 61 B2 19 8E 4E E5 4B 93 8F 5D DB A9",
+    "AD F1 AE 2E CB 0D FC F4 2D 46 6E 1D 97 E8 D1 E9",
+    "4D 37 A5 75 5E 83 9E AB 82 9D B9 1C E0 CD 49 89",
+    "01 B6 BD 58 24 A2 5F 38 78 99 15 90 50 B8 95 E4",
+    "D0 91 C7 CE ED 0F B4 6F A0 CC F0 02 4A 79 C3 DE",
+    "A3 EF EA 51 E6 6B 18 EC 1B 2C 80 F7 74 E7 FF 21",
+    "5A 6A 54 1E 41 31 92 35 C4 33 07 0A BA 7E 0E 34",
+    "88 B1 98 7C F3 3D 60 6C 7B CA D3 1F 32 65 04 28",
+    "64 BE 85 9B 2F 59 8A D7 B0 25 AC AF 12 03 E2 F2",
+]);
